refactor: Placement now uses a String internally
I got pretty far with just using a string slice, but everything started
to fall apart when I tried to implement From<Board> for Placement.

I could not return a new Placement because the lifetime of the internal
string slice could not out live the scope of the from function. I was
essentially trying to return a reference to something that was
immediately going to be discarded. There might have been some sort of
workaround, but I couldn't figure it out.