feat: implemented pseudo legal move generation
This commit includes a lot of new code (sorry for not committing
earlier)! I added a new State struct that encapsulates move generation.
The generation itself seems fine, but I am not sold on the
structure/organization of the code.

My concerns stem from the disconnect between FEN and Board. In
delveam/chess, FEN and Board were combined, but I decided to decouple
them in this project. This has been fine so far, but I am not sure what
to do now that I need a structure that depends on both FEN and Board.
For now State fills that role, but I am open for suggestions on
refactoring/improvements.