feat: implemented State::analyze(Color)
This method has also been derived from delveam/chess, but with some
baked in optimizations/improvements. With that said, this method is
still not fast; however, legal move generation is finally correct.

Overall, our approach to move generation is still quite naive — I expect
a lot of the internals and API design to change moving forward — but at
least we have a foundation to build upon.

For future reference here is the method this commit is based on:
https://github.com/delveam/chess/blob/2c1e8582b73cb59d71e14a50f671d4ccbac34b95/src/chess/gm.cpp#L1156