perf: refactored State to deal with applying moves itself
Previously Fen handled applying moves, but this lead to constantly
having to convert a Placement to a Board and then a Fen to a Board. As
you can imagine, this was very wasteful, and our performance suffered.

Now that State deals with applying moves directly, we no longer have to
worry about Placement/Fen conversions; State only interfaces with a
Board. In order to achieve this, additional properties were added to
State. The following changes put Fen in an awkward space, as State is
basically just Fen but without Placement, but we can worry about that
later :^)

After all of these changes our perft_kiwipete benchmark is down to just
about 16ms!