perf: reworked how making a move affects a Board
Previously, applying a move to a Board returned a new Board; this commit
makes it so applying a move mutates the current board. Moreover, applying
a move now handles en passant and castling.

Our previous approach was definitely more functional (and maybe even more
idiomatic), but the goal of this project is to run on the web;
unfortunately the codebase must ultimately defer to a WASM/JS API (not
idiomatic Rust).

Parallelism would have been easier with our previous approach; however,
WASM does not support threading in the first place! We are going to have
to work hard for our performance, and this commit is just the beginning.