perf: impl "Bulk-counting" perft optimization
At a depth of one, we can skip invoking make_move/unmake_move because
the total amount of legal moves is the perft value.

This is such a simple change, yet our perft benchmark saw an increase of
about 10%! After all of the recent perf commits, the perft_kiwipete
benchmark is down to about 10ms.

The optimization implemented in this commit was derived from the
"Bulk-counting" section from this article on the Chess Programming Wiki:

https://www.chessprogramming.org/Perft