refactor: `perft` no longer returns a Result
Now that State's analyze method can no longer fail*, perft no longer has
to return a Result.

Solidifying Fen's correctness really improved the ergonomics of this
project as a whole. Catching invalid state as soon as possible really
improves your code!

*The perft method can still panic; however, panicking is truly warranted
as a panic would mean the program as a whole is incorrect and
unrecoverable from the end user. With that said, as long as all of our
tests pass, this will never happen.