feat: implemented uci go command
This commit most notably includes a naive `minimax` implementation. In
order to get this function to work, a lot of utility functions were also
needed (e.g. `evaluation`, `analyze`).

The minimax implementation is not great; I expect a lot of its internals
to change moving forward. Moreover, the overall implementation/structure
of our engine is also going to change a lot. I think it is about time we
look into separating code into modules.

With all that said, this commit is a great start!