chore: addressed TODO comment
I have benchmarked an approach that does not use `SearchNode` and it was
2% slower than the current approach (I would also like to add that the
code was much cleaner). Given that `minimax` is a hot function, I do not
think we should make it worse (yeah it's only 2%, but if we want to go
up against Stockfish then we are going to need every little bit of extra
performance!).

For future reference, the approach mentioned above was derived from this
article:

https://www.chessprogramming.org/Principal_Variation