feat: impl Engine::make_sequence(State, &[Lan])
Added a function to apply a collection of `Lan`s to a given `State`. It
is worth noting that unlike `State::make_move(Lan)` this function
verifies that the moves are legal.

I am not sure if this function should belong to Engine, but we can
always move it later.