feat: quiescence search now returns SearchNode
Previously, it was possible for the `go depth <depth>` command to return
an incorrect Score::Mate(i8). The amount of turns was incorrect because
the move list's length (which is used to calculate the amount of turns)
was tied to the depth of the minimax search, not the length of the
minimax and quiescence search (which could have accidentally found
mate). Now that the quiescence search returns a `SearchNode`, the
complete move list can be derived.